requirements-minimal.txt
Flask==2.2.2
tensorflow-cpu==2.10.0
numpy==1.21.6
Pillow==9.5.0
gunicorn==20.1.0